feat: move banner-layout component to overrides
