+++ packages/gatsby-theme-docs/src/layouts/release-notes-detail.js
@@ -16,7 +16,7 @@ import LayoutFooter from './internals/layout-footer';
 import LayoutPageWrapper from './internals/layout-page-wrapper';
 import LayoutPage from './internals/layout-page';
 import LayoutPageHeader from './internals/layout-page-header';
-import LayoutPageHeaderSide from './internals/layout-page-header-side';
+import LayoutPageHeaderSide from '../overrides/layout-page-header-side';
 import LayoutPageContent from './internals/layout-page-content';
 import PageContentInset from './internals/page-content-inset';
 import { ReleaseNotePageTitle } from '../components/release-note-heading';
