+++ packages/gatsby-theme-docs/src/overrides/layout-page-header-side.js
