+++ websites/documentation/src/@commercetools-docs/gatsby-theme-docs/overrides/layout-page-header-side.js
@@ -1,0 +1,5 @@
+import styled from '@emotion/styled';
+
+const Nothing = styled.aside``;
+
+export default Nothing;
